python/pyproject.toml
[project]
name = "tino-daemon"
version = "0.1.0"
description = "Tino quantitative trading daemon - NautilusTrader gRPC wrapper"
requires-python = ">=3.10,<3.13"
dependencies = [
    "nautilus_trader>=1.200.0",
    "grpcio>=1.60.0",
    "grpcio-tools>=1.60.0",
    "grpcio-health-checking>=1.60.0",
    "grpcio-reflection>=1.60.0",
    "psutil>=5.9.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-asyncio>=0.21",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
